+++ CHANGELOG.md
@@ -37,6 +37,8 @@
       instead.
     * The `opAssembler()` function is deprecated; use the new `serializeOps()`
       function instead.
+    * The `mergingOpAssembler()` function is deprecated; use the new
+      `squashOps()` generator function with `serializeOps()` instead.
 
 ### Notable enhancements